publish-service/pyproject.toml
[project]
name = "cr8tor-publish-service"
version = "0.1.0"
description = "Python FastAPI app to serve as Publish Service for CR8TOR project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cr8tor",
    "databricks-sqlalchemy>=2.0.4",
    "dlt[duckdb,filesystem,postgres]>=1.5.0",
    "fastapi>=0.115.6",
    "obiba-opal>=5.3.0",
    "pydantic-settings>=2.7.1",
    "pymssql>=2.3.4",
    "pymysql>=1.1.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.1",
]

[tool.uv.sources]
cr8tor = { git = "https://github.com/lsc-sde-crates/cr8tor.git", branch = "main" }

[tool.pytest.ini_options]
testpaths = ["tests"]
